+++ src/wxMaximaFrame.cpp
@@ -963,7 +963,7 @@ void wxMaximaFrame::SetupMaximaMenu() {
 
   {
     wxMenuItem *it = new wxMenuItem(m_MaximaMenu, ToolBar::menu_restart_id,
-                                    _("&Restart Maxima\tCtrl+Alt+R"), _("Restart Maxima"),
+                                    _("&Restart Maxima\tShift+Ctrl+M"), _("Restart Maxima"),
                                     wxITEM_NORMAL);
     it->SetBitmap(GetWorksheet()->m_mainToolBar->GetRestartBitmap(
                                                                wxRendererNative::Get().GetCheckBoxSize(this)));